+++ app/src/main/java/com/ikm/inventoryqrscanner/activity/ListActivity.kt
@@ -68,7 +68,7 @@ class ListActivity : BaseActivity() {
                     number = document.data["number"].toString(),
                     product = document.data["product"].toString(),
                     expDate = document.data["expDate"] as Timestamp,
-                    amount = document.data["amount"].toString().toInt(),
+                    amount = document.data["amount"].toString(),
                     type = document.data["type"].toString(),
                     location = document.data["location"].toString(),
                     condition = document.data["condition"].toString(),
